Move combine metric join logic to SqlQueryPlanJoinBuilder
This code makes more sense in the join builder than the dataflow
to sql plan converter.
